--- a/Configuration/ProcessModifiers/python/ticl_superclustering_mustache_pf_cff.py
+++ b/Configuration/ProcessModifiers/python/ticl_superclustering_mustache_pf_cff.py
@@ -1,0 +1,4 @@
+import FWCore.ParameterSet.Config as cms
+
+# Modifier (to be applied on top of ticl_v5 modifier) to use the pre-TICLv5 superclustering code (that translates tracksters to PFClusters before running Mustache)
+ticl_superclustering_mustache_pf = cms.Modifier()