--- a/Configuration/ProcessModifiers/python/ticl_superclustering_mustache_ticl_cff.py
+++ b/Configuration/ProcessModifiers/python/ticl_superclustering_mustache_ticl_cff.py
@@ -1,0 +1,4 @@
+import FWCore.ParameterSet.Config as cms
+
+# Modifier (to be applied on top of ticl_v5 modifier) to run superclustering using Mustache inside TICL (with tracksters as inputs)
+ticl_superclustering_mustache_ticl = cms.Modifier()