--- a/RecoHGCal/TICL/interface/TracksterLinkingAlgoBase.h
+++ b/RecoHGCal/TICL/interface/TracksterLinkingAlgoBase.h
@@ -30,11 +30,22 @@ namespace edm {
   class EventSetup;
 }  // namespace edm
 
+namespace cms {
+  namespace Ort {
+    class ONNXRuntime;
+  }
+}  // namespace cms
+
 namespace ticl {
   class TracksterLinkingAlgoBase {
   public:
-    TracksterLinkingAlgoBase(const edm::ParameterSet& conf, edm::ConsumesCollector)
-        : algo_verbosity_(conf.getParameter<int>("algo_verbosity")) {}
+    /** \param conf the configuration of the plugin
+     * \param onnxRuntime the ONNXRuntime, if onnxModelPath was provided in plugin configuration (nullptr otherwise)
+    */
+    TracksterLinkingAlgoBase(const edm::ParameterSet& conf,
+                             edm::ConsumesCollector,
+                             cms::Ort::ONNXRuntime const* onnxRuntime = nullptr)
+        : algo_verbosity_(conf.getParameter<int>("algo_verbosity")), onnxRuntime_(onnxRuntime) {}
     virtual ~TracksterLinkingAlgoBase(){};
 
     struct Inputs {
@@ -62,10 +73,14 @@ namespace ticl {
                             const edm::ESHandle<MagneticField> bfieldH,
                             const edm::ESHandle<Propagator> propH) = 0;
 
+    // To be called by TracksterLinksProducer at the start of TracksterLinksProducer::produce. Subclasses can use this to store Event and EventSetup
+    virtual void setEvent(edm::Event& iEvent, edm::EventSetup const& iEventSetup){};
+
     static void fillPSetDescription(edm::ParameterSetDescription& desc) { desc.add<int>("algo_verbosity", 0); };
 
   protected:
     int algo_verbosity_;
+    cms::Ort::ONNXRuntime const* onnxRuntime_;
   };
 }  // namespace ticl