@cmsbuild 2024-06-27
cms-bot internal usage<!-- bot cache: {"commits":{"55f6f7f162a0be45b1ccdb5f4fbf4b0f2cd401ba":{"files":["DataFormats/HGCalReco/interface/Trackster.h","RecoHGCal/TICL/plugins/PatternRecognitionbyCA.cc","RecoHGCal/TICL/plugins/PatternRecognitionbyCLUE3D.cc","RecoHGCal/TICL/plugins/PatternRecognitionbyCLUE3D.h","RecoHGCal/TICL/plugins/PatternRecognitionbyFastJet.cc","RecoHGCal/TICL/plugins/PatternRecognitionbyPassthrough.cc","RecoHGCal/TICL/plugins/SimTrackstersProducer.cc","RecoHGCal/TICL/plugins/TICLCandidateProducer.cc","RecoHGCal/TICL/plugins/TracksterLinksProducer.cc","RecoHGCal/TICL/plugins/TrackstersMergeProducer.cc","RecoHGCal/TICL/plugins/TrackstersPCA.cc","RecoHGCal/TICL/plugins/TrackstersPCA.h","RecoHGCal/TICL/python/CLUE3DHighStep_cff.py"],"squashed":false,"time":1719397811},"e8b0e18a09d0b0291d2ea0a02ffb5a43e4b9f770":{"files":["CommonTools/RecoAlgos/interface/MultiVectorManager.h","Configuration/ProcessModifiers/python/ticl_superclustering_dnn_cff.py","Configuration/ProcessModifiers/python/ticl_superclustering_mustache_pf_cff.py","Configuration/ProcessModifiers/python/ticl_superclustering_mustache_ticl_cff.py","Configuration/PyReleaseValidation/README.md","Configuration/PyReleaseValidation/python/upgradeWorkflowComponents.py","RecoHGCal/Configuration/python/RecoHGCal_EventContent_cff.py","RecoHGCal/TICL/interface/SuperclusteringDNNInputs.h","RecoHGCal/TICL/interface/TracksterLinkingAlgoBase.h","RecoHGCal/TICL/plugins/BuildFile.xml","RecoHGCal/TICL/plugins/EGammaSuperclusterProducer.cc","RecoHGCal/TICL/plugins/SuperclusteringSampleDumper.cc","RecoHGCal/TICL/plugins/TICLCandidateProducer.cc","RecoHGCal/TICL/plugins/TICLDumper.cc","RecoHGCal/TICL/plugins/TracksterLinkingPassthrough.h","RecoHGCal/TICL/plugins/TracksterLinkingPluginFactory.cc","RecoHGCal/TICL/plugins/TracksterLinkingPluginFactory.h","RecoHGCal/TICL/plugins/TracksterLinkingbyFastJet.h","RecoHGCal/TICL/plugins/TracksterLinkingbySkeletons.cc","RecoHGCal/TICL/plugins/TracksterLinkingbySkeletons.h","RecoHGCal/TICL/plugins/TracksterLinkingbySuperClusteringDNN.cc","RecoHGCal/TICL/plugins/TracksterLinkingbySuperClusteringDNN.h","RecoHGCal/TICL/plugins/TracksterLinkingbySuperClusteringMustache.cc","RecoHGCal/TICL/plugins/TracksterLinkingbySuperClusteringMustache.h","RecoHGCal/TICL/plugins/TracksterLinksProducer.cc","RecoHGCal/TICL/python/CLUE3DHighStep_cff.py","RecoHGCal/TICL/python/customiseForTICLv5_cff.py","RecoHGCal/TICL/python/customiseTICLFromReco.py","RecoHGCal/TICL/python/iterativeTICL_cff.py","RecoHGCal/TICL/python/superclustering_cff.py","RecoHGCal/TICL/python/ticlDumper_cff.py","RecoHGCal/TICL/src/SuperclusteringDNNInputs.cc","SimCalorimetry/HGCalAssociatorProducers/python/TSToSimTSAssociation_cfi.py","Validation/Configuration/python/hgcalSimValid_cff.py","Validation/HGCalValidation/python/HGCalValidator_cfi.py"],"squashed":false,"time":1719493425}},"emoji":{},"last_seen_sha":"e8b0e18a09d0b0291d2ea0a02ffb5a43e4b9f770","signatures":{}} -->