@cmsbuild 2024-06-27
A new Pull Request was created by  @tcuisset for master.

It involves the following packages:

- CommonTools/RecoAlgos (**reconstruction**)
- Configuration/ProcessModifiers (**operations**)
- Configuration/PyReleaseValidation (**upgrade, pdmv**)
- DataFormats/HGCalReco (**upgrade, reconstruction**)
- RecoHGCal/Configuration (**upgrade, reconstruction**)
- RecoHGCal/TICL (**upgrade, reconstruction**)
- SimCalorimetry/HGCalAssociatorProducers (**upgrade, simulation**)
- Validation/Configuration (**simulation, dqm**)
- Validation/HGCalValidation (**dqm**)


@AdrianoDee, @antoniovagnerini, @antoniovilela, @civanch, @cmsbuild, @davidlange6, @fabiocos, @jfernan2, @kskovpen, @mandrenguyen, @mdhildreth, @miquork, @nothingface0, @rappoccio, @rvenditti, @srimanob, @subirsarkar, @sunilUIET, @syuvivida, @tjavaid can you please review it and eventually sign? Thanks.
@Martin-Grunewald, @abbiendi, @ahinzmann, @apsallid, @bsunanda, @cseez, @fabiocos, @felicepantaleo, @forthommel, @gkasieczka, @hatakeyamak, @jdolen, @jhgoh, @lecriste, @lgray, @makortel, @missirol, @mmusich, @pfs, @rappoccio, @rovere, @sameasy, @sethzenz, @slomeo, @sobhatta, @vandreev11, @youyingli this is something you requested to watch as well.
@antoniovilela, @rappoccio, @sextonkennedy you are the release manager for this.

cms-bot commands are listed <a href="http://cms-sw.github.io/cms-bot-cmssw-cmds.html">here</a>