@felicepantaleo 2024-06-27
test parameters:  

- workflow_opts= -w upgrade
- workflow = 29888.203,29688.203