@tcuisset 2024-07-24
Timing studies for the PR (using 100 SingleElectron events, PU=200), showing the the main producers in the electron chain (using TimingModule). (NB: the DNN is only enabled with TICLv5, which is currently off by default. So there would only be change for the *.203/204/205 workflows)

| Producer  | DNN (s) | TICL-mustache (s) | PF Mustache (baseline) (s) |
| ------------- | ------------- | ------------- | ------------- |
| ticlTracksterLinksSuperclustering | 0.16 | 0.0077| -|
| ecalDrivenElectronSeeds | 0.66 | 0.80| 0.11|
| electronCkfTrackCandidates | 0.57 |0.60| 0.05|
| electronGsfTracks | 0.37 |0.36| 0.05|
| Full event (tracker reco+TICL+Gsf electrons)| 29.1924| 29.3598 | 29.215|

The time taken by the electron chain increases with the DNN, but stays negligible compared to overall event reco time. This is expected to be reduced by the new PID (that would reduce the seeding of GsfTracks from pileup / hadronic tracksters)